requirements.txt
cryptography>=2.2.2
pyjwt>=1.6.4
paho-mqtt>=1.3.1